# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sales operations API. All routes except login require an authenticated employee session.
servers:
  - url: /api
paths:
...
  /sales:
    get:
      operationId: listSales
      responses:
        "200":
          description: Scoped daily reports
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/SalesReport"}}
    post:
      operationId: saveSales
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SalesInput"}
      responses:
        "201":
          description: Saved draft or submitted report
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SalesReport"}
  /attendance:
    get:
      operationId: listAttendance
      responses:
        "200":
          description: Scoped punch history
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Punch"}}
    post:
      operationId: punchAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PunchInput"}
      responses:
        "201":
          description: Recorded attempt
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Punch"}
...
components:
...
  schemas:
...
    SalesLine:
      type: object
      required: [productId, quantity, amount, returns]
      properties:
        productId: {type: integer}
        quantity: {type: integer}
        amount: {type: string}
        returns: {type: string}
    SalesInput:
      type: object
      required: [branchId, businessDate, lines, submitted]
      properties:
        branchId: {type: integer}
        businessDate: {type: string}
        lines: {type: array, items: {$ref: "#/components/schemas/SalesLine"}}
        submitted: {type: boolean}
        notes: {type: string}
    SalesReport:
      type: object
      required: [id, sellerId, branchId, businessDate, status, gross, returns, net, lines]
      properties:
        id: {type: integer}
        sellerId: {type: integer}
        branchId: {type: integer}
        businessDate: {type: string}
        status: {type: string}
        gross: {type: string}
        returns: {type: string}
        net: {type: string}
        lines: {type: array, items: {$ref: "#/components/schemas/SalesLine"}}
        notes: {type: ["string", "null"]}
    PunchInput:
      type: object
      required: [branchId, type, latitude, longitude, accuracy, capturedAt]
      properties:
        branchId: {type: integer}
        type: {type: string, enum: [IN, OUT]}
        latitude: {type: number}
        longitude: {type: number}
        accuracy: {type: number}
        capturedAt: {type: string}
    Punch:
      type: object
      required: [id, sellerId, branchId, type, status, receivedAt]
      properties:
        id: {type: integer}
        sellerId: {type: integer}
        branchId: {type: integer}
        type: {type: string}
        status: {type: string}
        receivedAt: {type: string}
        distanceMeters: {type: ["number", "null"]}
...